WTA #57: Renamed :project-list module to :search.
