--- project-list/src/main/java/com/jacob/wakatimeapp/projectlist/ui/ProjectListNavigator.kt
@@ -1,3 +1,0 @@
-package com.jacob.wakatimeapp.projectlist.ui
-
-interface ProjectListNavigator
