--- project-list/src/main/java/com/jacob/wakatimeapp/projectlist/ui/ProjectListScreen.kt
@@ -1,1 +1,0 @@
-package com.jacob.wakatimeapp.projectlist.ui
