+++ search/src/main/AndroidManifest.xml
