+++ search/src/main/java/com/jacob/wakatimeapp/search/data/network/SearchProjectAPI.kt
@@ -1,12 +1,12 @@
-package com.jacob.wakatimeapp.projectlist.data.network
+package com.jacob.wakatimeapp.search.data.network
 
-import com.jacob.wakatimeapp.projectlist.data.network.dto.ProjectListDTO
+import com.jacob.wakatimeapp.search.data.network.dto.ProjectListDTO
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Header
 import retrofit2.http.Query
 
-internal interface ProjectListAPI {
+internal interface SearchProjectAPI {
     @GET("/api/v1/users/current/projects")
     suspend fun getAllProjects(
         @Header("Authorization") token: String,
