+++ search/src/main/java/com/jacob/wakatimeapp/search/ui/SearchProjectNavigator.kt
@@ -1,0 +1,3 @@
+package com.jacob.wakatimeapp.search.ui
+
+interface SearchProjectNavigator
